Cleaning up code that produces compiler warnings. Removing some disused code.
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/map65@1009 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ loc.c
@@ -1,4 +1,4 @@
 int loc_(void *a)
 {
-  return a;
+  return (int)a;
 }
